+++ tests/test.rs
@@ -217,3 +217,55 @@ fn richtext_namespaced() {
                 Empty,
                 String("shared string\r\nLine 2\r\nLine 3".to_string())]]);
 }
+
+#[test]
+fn defined_names_xlsx() {
+    let path = format!("{}/tests/issues.xlsx", env!("CARGO_MANIFEST_DIR"));
+    let mut excel = Sheets::open(&path).expect("cannot open excel file");
+
+    let mut defined_names = excel.defined_names().unwrap().to_vec();
+    defined_names.sort();
+    assert_eq!(defined_names,
+               vec![("MyBrokenRange".to_string(), "Sheet1!#REF!".to_string()),
+                    ("MyDataTypes".to_string(), "datatypes!$A$1:$A$6".to_string()),
+                    ("OneRange".to_string(), "Sheet1!$A$1".to_string())]);
+}
+
+#[test]
+fn defined_names_xlsb() {
+    let path = format!("{}/tests/issues.xlsb", env!("CARGO_MANIFEST_DIR"));
+    let mut excel = Sheets::open(&path).expect("cannot open excel file");
+
+    let mut defined_names = excel.defined_names().unwrap().to_vec();
+    defined_names.sort();
+    assert_eq!(defined_names,
+               vec![("MyBrokenRange".to_string(), "Sheet1!#REF!".to_string()),
+                    ("MyDataTypes".to_string(), "datatypes!$A$1:$A$6".to_string()),
+                    ("OneRange".to_string(), "Sheet1!$A$1".to_string())]);
+}
+
+#[test]
+fn defined_names_xls() {
+    let path = format!("{}/tests/issues.xls", env!("CARGO_MANIFEST_DIR"));
+    let mut excel = Sheets::open(&path).expect("cannot open excel file");
+
+    let mut defined_names = excel.defined_names().unwrap().to_vec();
+    defined_names.sort();
+    assert_eq!(defined_names,
+               vec![("MyBrokenRange".to_string(), "Sheet1!#REF!".to_string()),
+                    ("MyDataTypes".to_string(), "datatypes!$A$1:$A$6".to_string()),
+                    ("OneRange".to_string(), "Sheet1!$A$1".to_string())]);
+}
+
+#[test]
+fn defined_names_ods() {
+    let path = format!("{}/tests/issues.ods", env!("CARGO_MANIFEST_DIR"));
+    let mut excel = Sheets::open(&path).expect("cannot open excel file");
+
+    let mut defined_names = excel.defined_names().unwrap().to_vec();
+    defined_names.sort();
+    assert_eq!(defined_names,
+               vec![("MyBrokenRange".to_string(), "of:=[Sheet1.#REF!]".to_string()),
+                    ("MyDataTypes".to_string(), "datatypes.$A$1:datatypes.$A$6".to_string()),
+                    ("OneRange".to_string(), "Sheet1.$A$1".to_string())]);
+}